requirements.txt
allure-pytest==2.13.5
pytest==8.1.1
requests==2.31.0
selene==2.0.0rc7
curlify==2.2.1
python-dotenv~=1.0.1
